+++ web/src/js/components/FlowTable.tsx
@@ -12,7 +12,7 @@ type FlowTableProps = {
     flows: Flow[];
     rowHeight: number;
     highlight: string;
-    selected: Flow;
+    selected: string[];
 };
 
 type FlowTableState = {
@@ -60,7 +60,8 @@ export class PureFlowTable extends React.Component<
         this.onViewportUpdate();
 
         const selectedNewFlow =
-            this.props.selected && this.props.selected !== prevProps.selected;
+            this.props.selected[0] &&
+            this.props.selected[0] !== prevProps.selected[0];
         if (selectedNewFlow) {
             const { rowHeight, flows, selected } = this.props;
             const viewport = this.viewport.current!;
@@ -68,7 +69,10 @@ export class PureFlowTable extends React.Component<
 
             const headHeight = head ? head.offsetHeight : 0;
 
-            const rowTop = flows.indexOf(selected) * rowHeight + headHeight;
+            const selectedFlowIndex = flows.findIndex(
+                (flow) => flow.id === selected[0],
+            );
+            const rowTop = selectedFlowIndex * rowHeight + headHeight;
             const rowBottom = rowTop + rowHeight;
 
             const viewportTop = viewport.scrollTop;
@@ -137,8 +141,8 @@ export class PureFlowTable extends React.Component<
                             <FlowRow
                                 key={flow.id}
                                 flow={flow}
-                                selected={flow === selected}
-                                highlighted={isHighlighted(flow)}
+                                isSelected={selected.includes(flow.id)}
+                                isHighlighted={isHighlighted(flow)}
                             />
                         ))}
                         <tr style={{ height: vScroll.paddingBottom }} />
@@ -152,5 +156,5 @@ export class PureFlowTable extends React.Component<
 export default connect((state: RootState) => ({
     flows: state.flows.view,
     highlight: state.flows.highlight,
-    selected: state.flows.byId[state.flows.selected[0]],
+    selected: state.flows.selected,
 }))(PureFlowTable);