+++ web/src/js/components/MainView.tsx
@@ -8,8 +8,10 @@ import Modes from "./Modes";
 import { Tab } from "../ducks/ui/tabs";
 
 export default function MainView() {
-    const hasSelection = useAppSelector(
-        (state) => !!state.flows.byId[state.flows.selected[0]],
+    const hasFlowSelected = useAppSelector(
+        (state) =>
+            state.flows.selected.length === 1 &&
+            !!state.flows.byId[state.flows.selected[0]],
     );
     const hasFlows = useAppSelector((state) => state.flows.list.length > 0);
     const currentTab = useAppSelector((state) => state.ui.tabs.current);
@@ -21,8 +23,12 @@ export default function MainView() {
             ) : (
                 <>
                     {hasFlows ? <FlowTable /> : <CaptureSetup />}
-                    {hasSelection && <Splitter key="splitter" />}
-                    {hasSelection && <FlowView key="flowDetails" />}
+                    {hasFlowSelected && (
+                        <>
+                            <Splitter key="splitter" />
+                            <FlowView key="flowDetails" />
+                        </>
+                    )}
                 </>
             )}
         </div>